docker_run: fix nonexistent directory in left side of -v.
Due to a bug in the sonic-management Makefile.work, it fails
to create a certain local directory. It then instructs docker
to map this nonexistent directory into a container.

This works because in the expected environment, Docker will
make up for the problem by creating the nonexistent directory.

In our environment, that cannot work: the docker we are using
is running as root, and cannot create that directory in the
NFS-mounted tree owned by us, and so docker run fails.

The bug in sonic-management's Makefile.work is that
$(mkdir -p ...) is used instead of $(shell mkdir -p ...).

* docker.sh (docker_run): If the left hand path of a -v is
inside our tree (i.e. $this_dir/ is a prefix of that path)
then perform a mkdir -p on that path to make sure it exists.
Then docker doesn't need to create it and can successfully
map it.