docker_run: assume short options take arg, like long ones.
* docker.sh (docker_run): in the --* case, we should also
handle -*, because docker has short options that take
arguments. (All those options that don't take arguments have
to be listed in a preceding case, not to end up here,
including common combinations of options like -it).
I realize that the -* pattern matches anything that --*
matches, but I want to leave the distinction visible.